+++ docs/changelog.rst
@@ -111,6 +111,7 @@ Other changes:
 * Refactored library.py file by splitting it into multiple modules within the
   beets/library directory.
 * Added a test to check that all plugins can be imported without errors.
+* Moved `beets/random.py` into `beetsplug/random.py` to cleanup core module.
 
 2.3.1 (May 14, 2025)
 --------------------